zero/backend/requirements.txt
Flask
flask-cors
flask-jwt-extended
pymongo
python-dotenv
bcrypt

